♻️ Decompose lib.rs into memory and composition modules (Task_5)
lib.rs reduces to crate docs, module declarations, and public
re-exports. The CharacterMemory facade moves to src/memory.rs; the
composition root (backend selection, factories, stats fallback,
embedder adapter glue) moves to src/composition.rs. Shared test fakes
dedupe into src/test_support.rs. Config flattens to one tree including
EmbeddingProviderSettings; errors flatten to a single file;
src/internal is removed entirely; mockall becomes a dev-dependency.

Local guardrail integration tests fail with the known machine-local
post-idle Qdrant gRPC stall; baseline HEAD fails identically (verified
via stash rerun), so CI remains the arbiter for those three tests.
