+++ src/config.rs
@@ -1,1 +1,7 @@
-pub mod settings;
+mod app_settings;
+mod embedding_provider_settings;
+
+pub use app_settings::{
+    GraphStoreMode, RetrievalStatsHealthFailMode, RetrievalStatsStoreMode, Settings,
+};
+pub(crate) use embedding_provider_settings::EmbeddingProviderSettings;
