+++ src/config/embedding_provider_settings.rs
