--- src/config/settings.rs
@@ -1,6 +1,0 @@
-mod app_settings;
-
-pub(crate) use crate::internal::config::settings::EmbeddingProviderSettings;
-pub use app_settings::{
-    GraphStoreMode, RetrievalStatsHealthFailMode, RetrievalStatsStoreMode, Settings,
-};
