--- src/internal.rs
@@ -1,2 +1,0 @@
-pub(crate) mod config;
-pub(crate) mod repositories;
