--- src/internal/config.rs
@@ -1,1 +1,0 @@
-pub(crate) mod settings;
