--- src/internal/config/settings.rs
@@ -1,3 +1,0 @@
-mod embedding_provider_settings;
-
-pub(crate) use embedding_provider_settings::EmbeddingProviderSettings;
