+++ src/usecases.rs
@@ -7,15 +7,8 @@ pub(crate) mod write_planning;
 
 pub(crate) use correct_forget::CorrectionForgetPipeline;
 pub(crate) use link::{admit_link, LinkAdmissionDecision, LinkAdmissionEvidence, LinkPipeline};
-pub(crate) use reconciliation::{
-    reconcile_graph_vector_stores, ReconciliationDiagnostic, ReconciliationDriftKind,
-    ReconciliationReport,
-};
 pub(crate) use remember::{
     RememberPipeline, RememberPipelineDraft, RememberPipelineOutcome, VectorIndexingFailure,
 };
 pub(crate) use retrieve::RetrievePipeline;
-pub(crate) use write_planning::{
-    WritePlanCommitValues, WritePlanValidationDecision, WritePlanValidationVerdict,
-    WritePlanValidator,
-};
+pub(crate) use write_planning::{WritePlanCommitValues, WritePlanValidator};
